Change init to initIt.
This is required for Swift because init is a keyword.

I presume that this isn't going to fly if it's broken the behavior for other
backends -- I expect we'll need to parameterize this, but just committing it
for now.

This comes from https://github.com/janyou/ANTLR-Swift-Target and is
marked Copyright (c) 2016 janyou on top of the BSD license and Copyrights
for Terence Parr and Sam Harwell derived from the original ANTLR source.