[Swift] Optimize bitmask tests.
This adds a computed mask to TestSetInline.Bitset, so that we can use
this during code generation.  It adds a corresponding BitsetRenderer
to SwiftTarget, and registers that with StringTemplate as an
AttributeRenderer.

This is then used in the Swift codegen template to remove the code
that called Utils.testBitLeftShiftArray.  This means that we have the
bitmask hardcoded in the parser, rather than passing an array of token
numbers and computing the mask each time we test against it.  It's not
clear how much this saves with an optimized build (Swift compiler doing
a great job?) but it saves something like a factor of 5 in parse time
with a debug build.