[Swift] Remove uses of "import Foundation".
We have very few dependencies on Foundation: UUID, NSUUID, and one use each
of InputStream, JSONSerialization, NSString, ProcessInfo, pthread_mutex_t
and fputs.

Make all of these explicit, and remove the other imports, so that we
don't accidentally use other parts.  In particular, String -> NSString
silent conversion performs a transcode in Swift 5, which is expensive and
so we need to avoid it on any hot paths.