[Swift] Improve ProfilingATNSimulator.
Replace use of Date().timeIntervalSince1970 with CACurrentMediaTime.
This is lower cost.

Fix the value that is put into timeInPrediction so that it is in
nanoseconds.  This is what is documented (carried over from the Java
runtime) but we were actually putting milliseconds in here.

Fix the defer block that decrements self.currentDecision.  This came over
from a finally block in Java, but it needs to be at the top of the method,
not the bottom, otherwise it will be bypassed in the event of an exception.

Change some vars to lets.