[Swift] Tidy-up around ATNDeserializer.
Make the ATNDeserializer.*factory methods static and private.

Remove ATNSimulator.edgeFactory.  This is deprecated in the Java runtime,
so we didn't need it in this port in the first place.

Don't force-unwrap the return value from stateFactory.  Returning nil
is meaningful here (it means that the given type was INVALID_TYPE).
The Java implementation handles this and proceeds to add the null state
to the ATN; we should do the same.

Move some helper functions from CommonUtil into ATNDeserializer, since
they are not common at all.