[Swift] Various fixes for BitSet.
Change the internals of BitSet to use UInt64, not Int64.  The original
implementation was apparently taken from OpenJDK's BitSet, but that uses
signed integers only because Java doesn't have unsigned integers.  Swift
does.  By using unsigned integers, we don't need our own
>>> operator (unsigned-right-bit-shift), because Swift's built-in >>
operator does the right thing when it is operating on unsigned integers.

Fix {previous,next}{Clear,Set}Bit and the toIndex versions of set, clear,
get, and flip.  The OpenJDK implementation uses
WORD_MASK = 0xffffffffffffffffL, but we were using WORD_MASK = Int64.max.
This is missing the top bit, so any operation using it was broken.

Note also that in Java (n >>> -1) is equivalent to (n >>> (-1 & 63)), but
in Swift the equivalent is not true.  OpenJDK's implementation relies on
this as a trick when calculating lastWordMask.  We were hiding this in
our >>> implementation, with a "% 64" thrown in there.  This is now explicit
in the lastWordMask initializations instead.

Remove the sizeIsSticky field.  This was written but never read.  It comes
from an optimization in the OpenJDK implementation, and is not used here.

Remove the serialVersionUID field.  This is only meaningful in Java.

This also adds some unit tests for BitSet.